# Generate optimized mmap data arrays for GAIA data.