I'm not sure what license this data is under, so we'll just ask
people to apply to NOAO Data Lab and generate it themselves until
we double-check.

Benchmarks for mmap stuff:

Time for reading pickle file: 1.9 s
Time per dataframe random id selection: 1.1e+03 ms
Time for reading mmap data and index: 0.00022 s
Time per random id selection (mmap arrays): 0.733137130737 ms

Obviously you can do a little better with pandas if you generate
an index. Also 0.77 ms is not fast given each light curve is just
a few points, but it should be sufficient not to be a training
pipeline bottleneck.